requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.0.4
beautifulsoup4==4.7.1
black==20.8b1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
click==7.1.2
colorama==0.4.0
dataclasses==0.8
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
django-appconf==1.0.3
django-ckeditor==5.8.0
django-el-pagination==3.2.4
django-haystack==2.8.1
django-imagekit==4.0.2
django-ipware==2.1.0
django-js-asset==1.2.2
django-summernote==0.8.11.6
django-xforwardedfor-middleware==2.0
djangorestframework==3.11.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysql==0.0.2
mysqlclient==1.4.6
oauthlib==3.1.0
pathspec==0.8.1
pilkit==2.0
Pillow==6.2.1
pylint==2.1.1
pylint-django==2.3.0
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2018.6
regex==2020.11.13
requests==2.21.0
requests-oauthlib==1.3.0
six==1.11.0
soupsieve==1.9
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.24.2
Whoosh==2.7.4
wrapt==1.10.11
